pyproject.toml
[project]
name = "tba-data-lake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.83",
    "duckdb>=1.5.1",
    "httpx>=0.28.1",
    "polars>=1.39.3",
    "python-dotenv>=1.2.2",
]
